add CheckLockMiddleware

+++ sparrow_cloud/middleware/lock_middleware.py
@@ -6,6 +6,9 @@ from sparrow_cloud.restclient import rest_client
 
 logger = logging.getLogger(__name__)
 
+SC_SPARROW_DISTRIBUTED_LOCK_SVC = get_cm_value("SC_SPARROW_DISTRIBUTED_LOCK_SVC")
+SC_SPARROW_DISTRIBUTED_LOCK_FRONT_API = get_cm_value("SC_SPARROW_DISTRIBUTED_LOCK_FRONT_API")
+
 class CheckLockMiddleware(MiddlewareMixin):
     def process_request(self, request):
         # key = request.headers.get("Sc-Lock")
@@ -13,15 +16,25 @@ class CheckLockMiddleware(MiddlewareMixin):
         # key不存在或者值为空，跳过锁检查，放行
         if not key:
             return
-        SC_SPARROW_DISTRIBUTED_LOCK_SVC = get_cm_value("SC_SPARROW_DISTRIBUTED_LOCK_SVC")
-        SC_SPARROW_DISTRIBUTED_LOCK_FRONT_API = get_cm_value("SC_SPARROW_DISTRIBUTED_LOCK_FRONT_API")
         try:
-            response = rest_client.delete(SC_SPARROW_DISTRIBUTED_LOCK_SVC, SC_SPARROW_DISTRIBUTED_LOCK_FRONT_API, json={"key":key})
-            res = response.get("result")
-            if res: # result不是0,则表示key存在且删除成功，只有在key存在的情况下表明是第一次提交，之后的提交会由于key不存在而禁止
+            response = rest_client.delete(SC_SPARROW_DISTRIBUTED_LOCK_SVC, SC_SPARROW_DISTRIBUTED_LOCK_FRONT_API, json={"key":key,"opt":"incr"})
+            res = response.get("code")
+            if res==0:
                 return
             else: # 锁不存在，可能已经被删除或者超时过期
                 return JsonResponse({"message":"重复提交，本次操作被禁止"}, status=403)
         except Exception as e:
+            # 发生异常时，直接放行
             logger.info("check front lock failed in lock middleware, message: {}".format(e.__str__()))
             return
+
+    def process_response(self, request, response):
+        key = request.META.get("HTTP_SC_LOCK")
+        if not key:
+            return response
+        # 业务处理正常，删除锁中的key
+        if response.status_code>=200 and response.status_code<300:
+            rest_client.delete(SC_SPARROW_DISTRIBUTED_LOCK_SVC, SC_SPARROW_DISTRIBUTED_LOCK_FRONT_API, json={"key":key})
+        else:#业务失败，重置锁，可以进行下一次请求
+            rest_client.put(SC_SPARROW_DISTRIBUTED_LOCK_SVC, SC_SPARROW_DISTRIBUTED_LOCK_FRONT_API, json={"key":key,"opt":"reset"})
+        return response
